Restructured React app
