+++ app/javascript/components/CourseNavBar/CourseNavBar.jsx
